made byte_reader/writer non-experimental, added convenience funs for mmapped files
